backend/requirements.txt
# Backend Dependencies

# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# NLP - spaCy (pre-built wheels)
spacy>=3.7.0
# Download model: python -m spacy download en_core_web_lg

# PII Detection - Presidio
presidio-analyzer>=2.2.30
presidio-anonymizer>=2.2.30

# Machine Learning (newer versions have better pre-built support)
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# LLM Integration
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
google-generativeai

# Database - MongoDB
pymongo>=4.6.0
motor>=3.3.0
dnspython>=2.5.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# CORS
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data Validation
email-validator>=2.1.0

# HTTP Requests
httpx>=0.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# PDF Generation
reportlab>=4.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
